+++ messages/fr/media.php
@@ -33,7 +33,7 @@ return [
     'ASSET_MODEL_CLASS_LABEL' => 'Modèle',
     'ASSET_MODEL_LABEL' => 'Enregistrement',
     'ASSET_SECTION_ASSET' => 'Ressource de section',
-    'ASSET_SUCCESS_CREATED' => 'La ressource a été créée.',
+    'ASSET_SUCCESS_CREATED' => 'La ressource « {name} » a été créée.',
     'ASSET_SUCCESS_DELETED' => 'La ressource a été supprimée.',
     'ASSET_SUCCESS_FILE_REPLACED' => 'Le fichier de la ressource a été remplacé.',
     'ASSET_SUCCESS_ORDERED' => 'L’ordre des ressources a été modifié.',